requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
astroid==2.2.5
atomicwrites==1.3.0
attrs==18.2.0
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==3.1.0
bs4==0.0.1
cachetools==4.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cloudpickle==1.3.0
colorama==0.4.1
coverage==4.5.2
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
Django==3.0.5
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
Flask==1.0.2
Flask-SQLAlchemy==2.4.4
future==0.18.2
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
gunicorn==20.0.4
gym==0.17.1
h5py==2.10.0
idna==2.8
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10
joblib==0.16.0
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==5.0.0
MouseInfo==0.1.3
mpmath==1.1.0
msgpack==0.6.2
mysql-connector-python==8.0.23
nbconvert==5.4.1
nbformat==4.4.0
nltk==3.4
nose==1.3.7
notebook==5.7.4
numpy==1.16.2
oauthlib==3.1.0
openpyxl==3.0.4
opt-einsum==3.1.0
pandocfilters==1.4.2
parso==0.3.4
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.8.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.11.2
psycopg2==2.8.6
py==1.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
pyglet==1.5.0
Pygments==2.3.1
pylint==2.3.1
PyMsgBox==1.0.8
pyparsing==2.3.1
pyperclip==1.7.0
PyRect==0.1.4
pyrsistent==0.14.11
PyScreeze==0.1.26
pytest==4.2.0
pytest-cov==2.6.1
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2019.3
pywhatkit==2.1
pywinpty==0.5.5
PyYAML==5.2
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
requests-oauthlib==1.3.0
responses==0.10.5
rsa==4.0
sapcai==4.0.0
scikit-learn==0.23.1
scipy==1.2.1
selenium==3.141.0
Send2Trash==1.5.0
singledispatch==3.4.0.3
six==1.13.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
sqlparse==0.3.1
sympy==1.3
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
threadpoolctl==2.1.0
traitlets==4.3.2
typed-ast==1.3.4
urllib3==1.24.1
virtualenv==16.7.9
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wikipedia==1.4.0
wrapt==1.11.2
